Improve comments, refactor a little bit
